style: 화면 구성 순서 수정
