+++ lib/screens/voicepocket/post_text_screen.dart
@@ -108,7 +108,6 @@ class _PostTextScreenState extends State<PostTextScreen> {
 
   @override
   Widget build(BuildContext context) {
-    double buttonsize = MediaQuery.of(context).size.width * 0.3;
     return Scaffold(
       drawer: const Drawer(),
       appBar: AppBar(
